RES-2499: Move continual learning framework to standalone package
