+++ environment.yml
@@ -19,6 +19,8 @@ dependencies:
 
       # Install nupic.research in development mode with dev tools extras enabled
       - -e ".[dev]"
+      
+      - -e packages/continual_learning
 
       # Install projects extra dependency
       # See `setup.cfg` for available extras.
