+++ nupic/research/frameworks/vernon/experiments/__init__.py
@@ -19,7 +19,6 @@
 # http://numenta.org/licenses/
 # ----------------------------------------------------------------------
 
-from .cl_experiment import *
 from .meta_cl_experiment import *
 from .self_supervised_experiment import *
 from .supervised_experiment import *
