+++ nupic/research/frameworks/vernon/mixins/__init__.py
@@ -26,7 +26,6 @@ from .context_signal import *
 from .cutmix import CutMix, CutMixKnowledgeDistillation
 from .delay_load_checkpoint import *
 from .dendrite_metrics import *
-from .ewc import ElasticWeightConsolidation
 from .export_model import ExportModel
 from .extra_validations_per_epoch import *
 from .gradient_metrics import GradientMetrics
@@ -43,19 +42,16 @@ from .maxup import MaxupPerSample, MaxupStandard
 from .multi_cycle_lr import MultiCycleLR
 from .noise_robustness_test import NoiseRobustnessTest
 from .oml import OnlineMetaLearning
-from .permuted_mnist_task_indices import *
 from .profile import Profile
 from .profile_autograd import ProfileAutograd
 from .prototype_context import *
 from .prune_low_magnitude import PruneLowMagnitude
 from .prune_low_snr import PruneLowSNRGlobal, PruneLowSNRLayers
-from .reduce_lr_after_task import ReduceLRAfterTask
 from .regularize_loss import RegularizeLoss
 from .report_max_accuracy import ReportMaxAccuracy
 from .representation_overlap import *
 from .rezero_weights import RezeroWeights
 from .save_final_checkpoint import SaveFinalCheckpoint
-from .si import *
 from .step_based_logging import *
 from .torch_profiler import *
 from .track_representation_sparsity import *
