+++ packages/continual_learning/nupic/research/frameworks/continual_learning/__init__.py
@@ -1,5 +1,5 @@
 #  Numenta Platform for Intelligent Computing (NuPIC)
-#  Copyright (C) 2020, Numenta, Inc.  Unless you have an agreement
+#  Copyright (C) 2019, Numenta, Inc.  Unless you have an agreement
 #  with Numenta, Inc., for a separate license for this software code, the
 #  following terms and conditions apply:
 #
@@ -17,7 +17,3 @@
 #
 #  http://numenta.org/licenses/
 #
-
-from .common import *
-from .hdf5_utils import *
-from .auto_augment import ImageNetPolicy
