+++ packages/continual_learning/nupic/research/frameworks/continual_learning/mixins/ewc.py
