+++ packages/continual_learning/setup.cfg
@@ -1,5 +1,6 @@
+# ------------------------------------------------------------------------------
 #  Numenta Platform for Intelligent Computing (NuPIC)
-#  Copyright (C) 2020, Numenta, Inc.  Unless you have an agreement
+#  Copyright (C) 2021, Numenta, Inc.  Unless you have an agreement
 #  with Numenta, Inc., for a separate license for this software code, the
 #  following terms and conditions apply:
 #
@@ -17,7 +18,17 @@
 #
 #  http://numenta.org/licenses/
 #
+# ------------------------------------------------------------------------------
+[metadata]
+name = nupic.research.continual_learning
+version = 0.0.1.dev0
 
-from .common import *
-from .hdf5_utils import *
-from .auto_augment import ImageNetPolicy
+[options]
+zip_safe = False
+packages = find_namespace:
+install_requires =
+    nupic.research
+
+[options.packages.find]
+include = nupic.research.*
+
