+++ packages/continual_learning/setup.py
@@ -1,5 +1,6 @@
+# ------------------------------------------------------------------------------
 #  Numenta Platform for Intelligent Computing (NuPIC)
-#  Copyright (C) 2020, Numenta, Inc.  Unless you have an agreement
+#  Copyright (C) 2021, Numenta, Inc.  Unless you have an agreement
 #  with Numenta, Inc., for a separate license for this software code, the
 #  following terms and conditions apply:
 #
@@ -17,7 +18,7 @@
 #
 #  http://numenta.org/licenses/
 #
+# ------------------------------------------------------------------------------
+from setuptools import setup
 
-from .common import *
-from .hdf5_utils import *
-from .auto_augment import ImageNetPolicy
+setup()
