+++ projects/dendrites/permutedMNIST/experiments/base.py
@@ -30,6 +30,7 @@ import ray.tune as tune
 import torch
 import torch.nn.functional as F
 
+from nupic.research.frameworks.continual_learning import mixins as cl_mixins
 from nupic.research.frameworks.dendrites import DendriticMLP
 from nupic.research.frameworks.dendrites.dendrite_cl_experiment import (
     DendriteContinualLearningExperiment,
@@ -40,7 +41,7 @@ from nupic.research.frameworks.vernon import mixins
 
 class PermutedMNISTExperiment(
     mixins.RezeroWeights,
-    mixins.PermutedMNISTTaskIndices,
+    cl_mixins.PermutedMNISTTaskIndices,
     DendriteContinualLearningExperiment,
 ):
     pass
