+++ projects/dendrites/permutedMNIST/experiments/hyperparameter_search.py
@@ -33,6 +33,7 @@ import ray.tune as tune
 import torch
 import torch.nn.functional as F
 
+from nupic.research.frameworks.continual_learning import mixins as cl_mixins
 from nupic.research.frameworks.dendrites import DendriticMLP
 from nupic.research.frameworks.dendrites.dendrite_cl_experiment import (
     DendriteContinualLearningExperiment,
@@ -47,6 +48,7 @@ class NbSegmentSearchExperiment(
     mixins.RezeroWeights,
     mixins.PrototypeContext,
     mixins.PermutedMNISTTaskIndices,
+    cl_mixins.PermutedMNISTTaskIndices,
     DendriteContinualLearningExperiment,
 ):
     pass
