+++ projects/dendrites/permutedMNIST/experiments/prototype.py
@@ -32,6 +32,7 @@ import ray.tune as tune
 import torch
 import torch.nn.functional as F
 
+from nupic.research.frameworks.continual_learning import mixins as cl_mixins
 from nupic.research.frameworks.dendrites import DendriticMLP
 from nupic.research.frameworks.dendrites.dendrite_cl_experiment import (
     DendriteContinualLearningExperiment,
@@ -44,7 +45,7 @@ from nupic.torch.modules import KWinners
 
 class PrototypeExperiment(mixins.RezeroWeights,
                           mixins.PrototypeContext,
-                          mixins.PermutedMNISTTaskIndices,
+                          cl_mixins.PermutedMNISTTaskIndices,
                           DendriteContinualLearningExperiment):
     pass
