+++ projects/dendrites/permutedMNIST/experiments/si_prototype.py
@@ -32,6 +32,7 @@ import ray.tune as tune
 import torch
 import torch.nn.functional as F
 
+from nupic.research.frameworks.continual_learning import mixins as cl_mixins
 from nupic.research.frameworks.dendrites import DendriticMLP
 from nupic.research.frameworks.dendrites.dendrite_cl_experiment import (
     DendriteContinualLearningExperiment,
@@ -43,7 +44,7 @@ from nupic.research.frameworks.vernon import mixins
 class SIPrototypeExperiment(mixins.SynapticIntelligence,
                             mixins.RezeroWeights,
                             mixins.PrototypeContext,
-                            mixins.PermutedMNISTTaskIndices,
+                            cl_mixins.PermutedMNISTTaskIndices,
                             DendriteContinualLearningExperiment):
     pass
