+++ projects/dendrites/permutedMNIST/experiments/sp_context.py
@@ -27,6 +27,7 @@ import ray.tune as tune
 import torch
 import torch.nn.functional as F
 
+from nupic.research.frameworks.continual_learning import mixins as cl_mixins
 from nupic.research.frameworks.dendrites import DendriticMLP
 from nupic.research.frameworks.dendrites.dendrite_cl_experiment import (
     DendriteContinualLearningExperiment,
@@ -37,7 +38,7 @@ from nupic.research.frameworks.vernon import mixins
 
 
 class SPExperiment(mixins.RezeroWeights,
-                   mixins.PermutedMNISTTaskIndices,
+                   cl_mixins.PermutedMNISTTaskIndices,
                    mixins.UpdateBoostStrength,
                    SpatialPoolerContext,
                    DendriteContinualLearningExperiment):
